Move elementsState logic to Hook
