+++ src/utils/hooks/useGetResponsesAndElementState.tsx
@@ -12,7 +12,7 @@ interface useLoadApplicationProps {
   serialNumber: string
 }
 
-const useGetResponsesByCode = (props: useLoadApplicationProps) => {
+const useGetResponsesAndElementState = (props: useLoadApplicationProps) => {
   const { serialNumber } = props
   const [responsesByCode, setResponsesByCode] = useState({})
   const [elementsExpressions, setElementsExpressions] = useState([])
@@ -65,11 +65,57 @@ const useGetResponsesByCode = (props: useLoadApplicationProps) => {
     setLoading(false)
   }, [data, apolloError])
 
+  useEffect(() => {
+    evaluateElementExpressions().then((result: any) => setElementsState(result))
+  }, [responsesByCode])
+
+  async function evaluateElementExpressions() {
+    const promiseArray: Promise<any>[] = []
+    elementsExpressions.forEach((element) => {
+      promiseArray.push(evaluateSingleElement(element))
+    })
+    const evaluatedElements = await Promise.all(promiseArray)
+    const elementsState: any = {}
+    evaluatedElements.forEach((element) => {
+      elementsState[element.code] = {
+        id: element.id,
+        category: element.category,
+        isEditable: element.isEditable,
+        isRequired: element.isRequired,
+        isVisible: element.isVisible,
+        // isValid: element.isValid,
+      }
+    })
+    return elementsState
+  }
+
+  async function evaluateSingleElement(element: any) {
+    const evaluationParameters = {
+      objects: [responsesByCode], // TO-DO: Also send user/org objects etc.
+      // graphQLConnection: TO-DO
+    }
+    const isEditable = evaluateExpression(element.isEditable)
+    const isRequired = evaluateExpression(element.isRequired)
+    const isVisible = evaluateExpression(element.visibilityCondition)
+    // const isValid = evaluateExpression(element.validation)
+    const results = await Promise.all([isEditable, isRequired, isVisible])
+    const evaluatedElement = {
+      code: element.code,
+      id: element.id,
+      category: element.category,
+      isEditable: results[0],
+      isRequired: results[1],
+      isVisible: results[2],
+      // isValid: results[3],
+    }
+    return evaluatedElement
+  }
+
   return {
     error,
     loading,
     responsesByCode,
-    elementsExpressions,
+    elementsState,
   }
 }
 
@@ -81,4 +127,4 @@ const useGetResponsesByCode = (props: useLoadApplicationProps) => {
 //   }
 //   return null
 // }
-export default useGetResponsesByCode
+export default useGetResponsesAndElementState
